[PPPSYS-43233] removed unused import in layout
