--- example-v13/app/layout.tsx
@@ -1,27 +1,0 @@
-'use client'
-
-import './globals.css'
-import type { Metadata } from 'next'
-import { Inter } from 'next/font/google'
-import PiwikProProvider from "@piwikpro/next-piwik-pro";
-
-const inter = Inter({ subsets: ['latin'] })
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <html lang="en">
-      <body className={inter.className}>
-        <PiwikProProvider
-          containerUrl={process.env.NEXT_PUBLIC_CONTAINER_URL}
-          containerId={process.env.NEXT_PUBLIC_CONTAINER_ID}
-        >
-          {children}
-        </PiwikProProvider>
-      </body>
-    </html>
-  )
-}
